Zelenova, week03, turtle lab 2-14
