+++ week03/zelenova_week03/zelenova_turt2.py
@@ -1,0 +1,16 @@
+import turtle
+
+#2 Следующая программа рисует букву S:
+turtle.shape('turtle')
+turtle.forward(50)
+turtle.left(90)
+turtle.forward(50)
+turtle.left(90)
+turtle.forward(50)
+turtle.right(90)
+turtle.forward(50)
+turtle.right(90)
+turtle.forward(50)
+
+
+
