+++ week03/zelenova_week03/zelenova_turt3.py
@@ -1,0 +1,12 @@
+#3 Квадрат.
+import turtle
+turtle.shape('turtle')
+turtle.forward(100)
+turtle.left(90)
+turtle.forward(100)
+turtle.left(90)
+turtle.forward(100)
+turtle.left(90)
+turtle.forward(100)
+turtle.left(90)
+
