+++ week03/zelenova_week03/zelenova_turt8.py
@@ -1,0 +1,9 @@
+#8 Квадратная спираль
+import turtle
+step=10 #сторона наименьшего квадрата
+n=100 #количество граней квадратов,пройденных черепашкой в спирале
+turtle.shape('turtle')
+for i in range(n):
+     turtle.forward(step)
+     turtle.left(90)
+     step+=5
